requirements.txt
asgiref==3.8.1
Babel==2.14.0
backports.zoneinfo==0.2.1
Django==4.2.13
django-bootstrap5==23.3
django-phonenumber-field==7.2.0
djangorestframework==3.15.1
mysqlclient==2.2.0
phonenumbers==8.13.26
Pillow==10.1.0
python-dotenv==1.0.0
pytz==2024.1
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
whitenoise==6.6.0
